Still trying to make work the sound in the example, 19

+++ kleine/lib/sound/sound_player.py
@@ -18,7 +18,7 @@ class SoundPlayer(PyXavi):
         "name": "wm8960soundcard",
         "channels": 1,
         "sample_rate": 44100,
-        "dtype": "int16",
+        "dtype": "float32",
         "blocksize": 0
     })
